+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/chaostoolkit/chaostoolkit-addons/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/chaostoolkit/chaostoolkit-addons/compare/0.3.0...HEAD
+
+## [0.3.0][]
+
+[0.3.0]: https://github.com/chaostoolkit/chaostoolkit-addons/compare/0.2.0...0.3.0
 
 ### Added